- bitrise.yml update - clean (+1 squashed commit) Squashed commits: [48d841d] - Webhook_URL set to secret. - config.go file's content moved to main.go
