+++ bitrise.yml
@@ -1,22 +1,22 @@
-format_version: 1.1.0
+format_version: 5
 default_step_lib_source: https://github.com/bitrise-io/bitrise-steplib.git
+
 app:
   envs:
-  - STEP_VERSION: 2.7.2
-  - SLACK_WEBHOOK_URL: $SLACK_WEBHOOK_URL
-  - SLACK_CHANNEL: $SLACK_CHANNEL
+  - BITRISE_STEP_ID: slack
+  - BITRISE_STEP_VERSION: 2.7.2
+  - BITRISE_STEP_GIT_CLONE_URL: https://github.com/bitrise-io/steps-slack-message.git
+  - MY_STEPLIB_REPO_FORK_GIT_URL: $MY_STEPLIB_REPO_FORK_GIT_URL
+
 workflows:
   test:
     before_run:
-    - go-tests
+    - audit-this-step
     steps:
-    - script:
-        title: go test
-        inputs:
-        - content: |
-            #!/bin/bash
-            set -ex
-            go test ./...
+    - go-list:
+    - golint:
+    - errcheck:
+    - go-test:
     - path::./:
         title: On Success
         is_skippable: false
@@ -111,64 +111,8 @@ workflows:
     - path::./:
         is_skippable: false
 
-  go-tests:
-    steps:
-    - script:
-        title: Install required testing tools
-        inputs:
-        - content: |-
-            #!/bin/bash
-            set -ex
-
-            # Check for unhandled errors
-            go get -u -v github.com/kisielk/errcheck
-
-            # Go lint
-            go get -u -v github.com/golang/lint/golint
-    - script:
-        title: Export go files to test
-        inputs:
-        - content: |-
-            #!/bin/bash
-            set -ex
-
-            no_vendor_paths="$(go list ./... | grep -v vendor)"
-            envman add --key GOLIST_WITHOUT_VENDOR --value "$no_vendor_paths"
-    - script:
-        title: Err check
-        inputs:
-        - content: |-
-            #!/bin/bash
-            set -ex
-
-            errcheck -asserts=true -blank=true $GOLIST_WITHOUT_VENDOR
-    - script:
-        title: Go lint
-        inputs:
-        - content: |-
-            #!/bin/bash
-            set -ex
-
-            while read -r line; do
-              echo "-> Linting: $line"
-              golint_out="$(golint $line)"
-              if [[ "${golint_out}" != "" ]] ; then
-                echo "=> Golint issues found:"
-                echo "${golint_out}"
-                exit 1
-              fi
-            done <<< "$GOLIST_WITHOUT_VENDOR"
-    - script:
-        title: Go test
-        inputs:
-        - content: |-
-            #!/bin/bash
-            set -ex
-
-            go test $(go list ./... | grep -v vendor)
   # ----------------------------------------------------------------
   # --- Utility workflows
-  # ----------------------------------------------------------------
   dep-update:
     title: Dep update
     description: |
@@ -183,8 +127,6 @@ workflows:
             go get -u -v github.com/golang/dep/cmd/dep
             dep ensure -v
             dep ensure -v -update
-
-  # ----------------------------------------------------------------
   # --- workflow to create Release version
   create-release:
     steps:
@@ -192,14 +134,14 @@ workflows:
         title:
         inputs:
         - content: |
-            #!/usr/bin/env bash
+            #!/bin/bash
             set -ex
             go get -u github.com/bitrise-tools/releaseman
             export CI=true
-            releaseman create --version "$STEP_VERSION"
-
+            releaseman create --version "$BITRISE_STEP_VERSION"
+            
   # ----------------------------------------------------------------
-  # --- workflow to Share this step into a Step Library
+  # --- workflows to Share this step into a Step Library
   audit-this-step:
     steps:
     - script:
@@ -209,41 +151,36 @@ workflows:
             set -ex
             stepman audit --step-yml ./step.yml
   share-this-step:
+    envs:
+      # if you want to share this step into a StepLib
+      - MY_STEPLIB_REPO_FORK_GIT_URL: $MY_STEPLIB_REPO_FORK_GIT_URL
+      - BITRISE_STEP_ID: $BITRISE_STEP_ID
+      - BITRISE_STEP_VERSION: $BITRISE_STEP_VERSION
+      - BITRISE_STEP_GIT_CLONE_URL: $BITRISE_STEP_GIT_CLONE_URL
     description: |-
       If this is the first time you try to share a Step you should
       first call: $ bitrise share
-
       This will print you a guide, and information about how Step sharing
       works. Please read it at least once!
-
       As noted in the Step sharing guide you'll have to fork the
       StepLib you want to share this step into. Once you're done with forking
       the repository you should set your own fork's git clone URL
       in the `.bitrise.secrets.yml` file, or here in the `envs` section,
-      as the value of the `MY_STEPLIB_REPO_FORK_GIT_URL` environment.
-
+      as the value of the `MY_STEPLIB_REPO_FORK_GIT_URL` environment. 
       You're now ready to share this Step, just make sure that
-      the `STEP_ID_IN_STEPLIB` and `STEP_GIT_VERION_TAG_TO_SHARE`
+      the `BITRISE_STEP_ID` and `BITRISE_STEP_VERSION`
       environments are set to the desired values!
-
       To share this Step into a StepLib you can just run: $ bitrise run share-this-step
-
       Once it finishes the only thing left is to actually create a Pull Request,
       the way described in the guide printed at the end of the process.
-    envs:
-    - MY_STEPLIB_REPO_FORK_GIT_URL: $MY_STEPLIB_REPO_FORK_GIT_URL
-    - STEP_ID_IN_STEPLIB: slack
-    - STEP_GIT_VERION_TAG_TO_SHARE: $STEP_VERSION
-    - STEP_GIT_CLONE_URL: https://github.com/bitrise-io/steps-slack-message.git
+    before_run:
+    - audit-this-step
     steps:
     - script:
         inputs:
         - content: |-
             #!/bin/bash
-            set -e
-            set -x
-
-            bitrise share start -c ${MY_STEPLIB_REPO_FORK_GIT_URL}
-            bitrise share create --stepid ${STEP_ID_IN_STEPLIB} --tag ${STEP_GIT_VERION_TAG_TO_SHARE} --git ${STEP_GIT_CLONE_URL}
+            set -ex
+            bitrise share start -c "${MY_STEPLIB_REPO_FORK_GIT_URL}"
+            bitrise share create --stepid "${BITRISE_STEP_ID}" --tag "${BITRISE_STEP_VERSION}" --git "${BITRISE_STEP_GIT_CLONE_URL}"
             bitrise share finish
-
